feat: Refactor EditorShell component and extract starter flow logic
- Removed unused imports and code related to project management and restore points from EditorShell.
- Introduced useEditorProjects hook to manage project and restore point logic.
- Created useEditorHints hook to handle hint dismissals and related state.
- Moved starter steps and actions into a new StarterFlow component for better organization.
- Updated EditorShell to utilize the new hooks and components, improving readability and maintainability.
